% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.R
\name{chain_stack}
\alias{chain_stack}
\title{Functions implementing stacking computation
adapted from codes by Yuling Yao
https://github.com/yao-yl/Multimodal-stacking-code}
\usage{
chain_stack(
  log_lik_mat,
  lambda = 1.0001,
  stack_iter = 1e+05,
  seed = 1,
  nsample = NULL,
  print_progress = TRUE
)
}
\arguments{
\item{log_lik_mat}{nSample x nChain x n log liklihood array.}

\item{lambda}{Dirichlet prior}

\item{stack_iter}{max iteration of optimization}

\item{seed}{random seed used in generating resampled draws}

\item{nsample}{number of samples to take}

\item{print_progress}{whether to print the progress message}
}
\description{
Functions implementing stacking computation
adapted from codes by Yuling Yao
https://github.com/yao-yl/Multimodal-stacking-code
}
\examples{

set.seed(12345)
loglik <- array(NA, dim = c(1000, 6, 20))
for(i in 1:20){
		lik <- runif(1, 0.1, 10)
		for(j in 1:6){
			loglik[, j, i] <- log(lik + runif(1000, -0.01, 0.01))
		}
}
stacked <- chain_stack(log_lik_mat = loglik, seed = 1, nsample = 500)
stacked$draws

}
