% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignment.R
\name{get_assignment}
\alias{get_assignment}
\title{function to extract the most likely cause of death from posterior draws}
\usage{
get_assignment(draws, Burn_in = NULL)
}
\arguments{
\item{draws}{nSample by n matrix}

\item{Burn_in}{number of iterations to discard}
}
\description{
function to extract the most likely cause of death from posterior draws
}
\examples{
\dontrun{
data(simLCM)
out.train <- LCVA.train(X = simLCM$X_train, 
				Y = simLCM$Y_train, Domain = simLCM$G_train, 
             causes.table = simLCM$causes.table, 
             domains.table = simLCM$domains.table,
           K = 5, model = "M", Nitr = 400, nchain = 5, seed = 1234)
out <- LCVA.pred(fit = out.train, X_test = simLCM$X_test,  
                 model = "D", 
                 alpha_pi = 1, alpha_eta = .1, 
                 Burn_in_train = 200, Nitr = 200)
Yhat <- get_assignment(out$Y_test, Burn_in = 100)
sum(simLCM$Y_test == Yhat) / length(Yhat)
}
}
