% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcm_pred}
\alias{lcm_pred}
\title{Internal function to predict with the nested latent class model}
\usage{
lcm_pred(
  X_test,
  Y_test,
  Group_test,
  config_train,
  N_test,
  S,
  C,
  K,
  G,
  itr_draws,
  alpha_pi_vec,
  alpha_eta,
  a_omega,
  b_omega,
  lambda_fit,
  phi_fit,
  pi_fit,
  pi_init,
  Nitr,
  similarity
)
}
\arguments{
\item{X_test}{a n by p matrix of the symptoms with 0 being absent, 1 being present, and NA being missing.}

\item{Y_test}{a vector of length n of the causes-of-death, coded into 0 to C, where C is the total number of all cause and 0 indicates unknown cause of death.}

\item{Group_test}{a vector of length n of the domain indicators, coded into 0 to G, where G is the total number of all domains and 0 indicates a new target domain.}

\item{config_train}{a matrix of counts for all domain-cause combinations.}

\item{N_test}{number of deaths to assign a cause to.}

\item{S}{number of symptoms.}

\item{C}{number of causes.}

\item{K}{number of latent classes within each cause-of-death.}

\item{G}{number of training domains.}

\item{itr_draws}{vector of iteration indices to use from the training posterior draws.}

\item{alpha_pi_vec}{vector of the concentration parameters for the target domain CSMF.}

\item{alpha_eta}{concentration parameter for the domain or domain-cause mixture prior. Only used for multi-domain models.}

\item{a_omega}{Shape parameter of the gamma distribution for the omega_c parameter in the stick-breaking prior.}

\item{b_omega}{Rate parameter of the gamma distribution for the omega_c parameter in the stick-breaking prior.}

\item{lambda_fit}{posterior draws of the training mix weights.}

\item{phi_fit}{posterior draws of the response probabilities.}

\item{pi_fit}{posterior draws of the training CSMF.}

\item{pi_init}{initial values of the target CSMF.}

\item{Nitr}{number of iterations to run in each MCMC chain.}

\item{similarity}{shrinkage model for the testing domain mixing weights. Currently not used.}
}
\description{
Internal function to predict with the nested latent class model
}
\examples{
message("See ?LCVA.pred")
}
