% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_train.R
\name{LCVA.train}
\alias{LCVA.train}
\title{Fit the nested latent class model on training data}
\usage{
LCVA.train(
  X,
  Y,
  Domain,
  K,
  model = NULL,
  Nitr = 4000,
  nchain = 1,
  seed = NULL,
  causes.table = NULL,
  domains.table = NULL,
  alpha_pi = 1,
  a_omega = 1,
  b_omega = 1,
  nu_phi = 1,
  nu_tau = 1,
  a_base = 1,
  b_base = 1
)
}
\arguments{
\item{X}{a n by p matrix of the symptoms with 0 being absent, 1 being present, and NA being missing.}

\item{Y}{a vector of length n of the causes-of-death. It can be one of the two formats: (1) a vector of cause-of-death label; (2) a numeric vector taking values from 1 to C, where C is the total number of all cause.}

\item{Domain}{a vector of length n of the domain indicators, coded into 1 to G, where G is the total number of all domains.}

\item{K}{number of latent classes within each cause-of-death}

\item{model}{the model to be fit. The current choice of models include Single-domain model with constant mixing weights ("S"), Single domain model with new mixing weights ("SN"), Multi-domain model with new mixing weights ("MN"), Multi-domain model with domain-level mixture ("MD"), Multi-domain model with domain-cause level mixture "MDC".}

\item{Nitr}{number of iterations to run in each MCMC chain.}

\item{nchain}{number of MCMC chains to run.}

\item{seed}{a vector of seeds to be used for each MCMC chain.}

\item{causes.table}{a vector of causes-of-death labels.}

\item{domains.table}{a vector of domains labels.}

\item{alpha_pi}{Concentration parameter for the training domain CSMF prior. Default to be 1}

\item{a_omega}{Shape parameter of the gamma distribution for the omega_c parameter in the stick-breaking prior. Can be either a scalar or a C by p matrix. Default to be 1}

\item{b_omega}{Rate parameter of the gamma distribution for the omega_c parameter in the stick-breaking prior. Can be either a scalar or a C by p matrix.  Default to be 1}

\item{nu_phi}{Shape2 parameter of the beta distribution for the class-dependent response probabilities. Default to be 1}

\item{nu_tau}{Shape2 parameter of the sparsity level in the response probabilities. Default to be 1}

\item{a_base}{Shape1 parameter of the gamma prior for the baseline response probabilities. Default to be 1}

\item{b_base}{Shape2 parameter of the gamma prior for the baseline response probabilities. Default to be 1}
}
\value{
a fitted object in the class of LCVA
}
\description{
Fit the nested latent class model on training data
}
\examples{
\dontrun{
data(simLCM)
out.train <- LCVA.train(X = simLCM$X_train, Y = simLCM$Y_train, 
						Domain = simLCM$G_train, 
			   			causes.table = simLCM$causes.table, 
					    domains.table = simLCM$domains.table,
						K = 5, model = "MD", Nitr = 400, nchain = 1, 
						seed = 1234)
summary(out.train)
}
}
