% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logsumexp}
\alias{logsumexp}
\title{Calculating the log of sum of exponentiation vector}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
log(sum(exp(x[1]) + exp(x[2]) + ...))
}
\description{
Calculating the log of sum of exponentiation vector
}
\examples{
logsumexp(c(1, 2, 3))
log(sum(exp(c(1, 2, 3))))
}
