% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_Dirichlet}
\alias{sample_Dirichlet}
\title{Sample from the Dirichlet distribution}
\usage{
sample_Dirichlet(a)
}
\arguments{
\item{a}{vector of Dirichlet concentration parameter}
}
\value{
a sample from Dirichlet(a).
}
\description{
Sample from the Dirichlet distribution
}
\examples{
alpha <- c(1, 2, 3, 4)
x <- matrix(0, 10000, 4)
for(i in 1:10000){
  x[i, ] <- sample_Dirichlet(alpha) 
}
round(apply(x, 2, mean), 2)
}
