% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_log_prob}
\alias{sample_log_prob}
\title{Sample from log of unnormalized probability vector}
\usage{
sample_log_prob(x)
}
\arguments{
\item{x}{a vector of log of unnormalized probability vector}
}
\value{
a sample from 0 to length(x)-1, with probabilities exp(x)/sum(exp(x))
}
\description{
Sample from log of unnormalized probability vector
}
\examples{
x <- rep(0, 10000)
prob <- c(1, 2, 3, 4)
logprob <- log(prob) 
for(i in 1:10000){
  x[i] <- sample_log_prob(logprob) 
}
round(table(x) / 10000, 2)
}
