% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_log_prob_matrix}
\alias{sample_log_prob_matrix}
\title{Sample from log of unnormalized probability matrix}
\usage{
sample_log_prob_matrix(x)
}
\arguments{
\item{x}{a matrix where each cell contains the log of a unnormalized probability matrix.}
}
\value{
the row and column index of a sample of the cells in the matrix. Index starting from 0.
}
\description{
Sample from log of unnormalized probability matrix
}
\examples{
x <- matrix(0, 10000, 2)
prob <- cbind(c(1, 2, 3, 4), c(4, 3, 2, 1))
logprob <- log(prob) 
for(i in 1:10000){
  x[i, ] <- sample_log_prob_matrix(logprob) 
}
round(table(paste(x[, 1], x[, 2], sep = "-")) / 10000, 2)
prob / sum(prob)
}
