% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_log_prob_matrix_col}
\alias{sample_log_prob_matrix_col}
\title{Sample from a column of the log of unnormalized probability matrix}
\usage{
sample_log_prob_matrix_col(x, id)
}
\arguments{
\item{x}{a matrix where each column contains the log of a unnormalized probability vector}

\item{id}{column index, starting from 0.}
}
\value{
a sample from 0 to length(x)-1, with probabilities exp(x[, id + 1])/sum(exp(x, id + 1))
}
\description{
Sample from a column of the log of unnormalized probability matrix
}
\examples{
x <- rep(0, 10000)
prob <- cbind(c(1, 2, 3, 4), c(4, 3, 2, 1))
logprob <- log(prob) 
for(i in 1:10000){
  x[i] <- sample_log_prob_matrix_col(logprob, 0) 
}
table(x) / 10000
for(i in 1:10000){
  x[i] <- sample_log_prob_matrix_col(logprob, 1) 
}
round(table(x) / 10000, 2)
}
