% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_prob}
\alias{sample_prob}
\title{Sample from a probability vector}
\usage{
sample_prob(x)
}
\arguments{
\item{x}{a vector of a probability vector}
}
\value{
a sample from 0 to length(x)-1, with probabilities x / sum(x).
}
\description{
Sample from a probability vector
}
\examples{
x <- rep(0, 10000)
prob <- c(.1, .2, .3, .4)
for(i in 1:10000){
  x[i] <- sample_prob(prob) 
}
round(table(x) / 10000, 2)
}
