% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLCM.R
\docType{data}
\name{simLCM}
\alias{simLCM}
\title{Simulated VA dataset from the latent class model.}
\format{
A list of data objects.
}
\usage{
data(simLCM)
}
\description{
A small simulated dataset with 1000 labeled deaths from 5 training domains and 1000 unlabeled death from a target domain. Data simulated to be have 10 causes and 50 symptoms.
}
\keyword{datasets}
