% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_pred.R
\name{summary.LCVA.pred}
\alias{summary.LCVA.pred}
\title{Summary method for the fitted model from \code{LCVA.pred}.}
\usage{
\method{summary}{LCVA.pred}(object, CI = 0.95, Burn_in = NULL, top = 5, ...)
}
\arguments{
\item{object}{output from \code{\link{LCVA.pred}}}

\item{CI}{credible interval to show in the summary}

\item{Burn_in}{number of initial iterations to discard as burn-in in the prediction stage.}

\item{top}{number of top causes to display.}

\item{...}{not used}
}
\description{
This function is the summary method for class \code{LCVA.pred}.
}
\examples{
\dontrun{
data(simLCM)
out.train <- LCVA.train(X = simLCM$X_train, Y = simLCM$Y_train, Domain = simLCM$G_train, 
             causes.table = simLCM$causes.table, 
             domains.table = simLCM$domains.table,
           K = 5, model = "MD", Nitr = 400, nchain = 1, seed = 1234)
out <- LCVA.pred(fit = out.train, X_test = simLCM$X_test,  
                 alpha_pi = 1, alpha_eta = .1, 
                 Burn_in_train = 200, Nitr = 200)
summary(out)
# Compare with the truth
sort(table(simLCM$Y_test)/length(simLCM$Y_test), dec = TRUE)
}
}
\seealso{
\code{\link{LCVA.pred}}
}
