% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_train.R
\name{summary.LCVA}
\alias{summary.LCVA}
\title{Summary method for the fitted model from \code{LCVA.train}.}
\usage{
\method{summary}{LCVA}(object, ...)
}
\arguments{
\item{object}{output from \code{\link{LCVA.train}}}

\item{...}{not used}
}
\description{
This function is the summary method for class \code{LCVA}.
}
\examples{
\dontrun{
data(simLCM)
out.train <- LCVA.train(X = simLCM$X_train, Y = simLCM$Y_train, 
						Domain = simLCM$G_train, 
			   			causes.table = simLCM$causes.table, 
					    domains.table = simLCM$domains.table,
						K = 5, model = "MD", Nitr = 400, nchain = 1, 
						seed = 1234)
summary(out.train)
}
}
\seealso{
\code{\link{LCVA.train}}
}
